Clear up DataFrame / AbstractDataFrame signatures
Fleshes out a current (transient) interface for AbstractDataFrames,
loosens signatures of methods that work for AbstractDataFrames,
tightens signatures of methods that don't.

Shuffles code around accordingly -- it is not meant to be permanent,
but apart from less broken methods, it is much clearer what implicit
interfaces exist, and, for me at least, what changes would need to
happen to add different types of data frames. Longer term, perhaps
DataFrame and some of its methods belong to an abstract subtype of
AbstractDataFrame.