Port from Nullable to Nulls
Keep Nullable in cases where it's used to represent "software engineer's nulls",
i.e. not missing values in data, since the replacement for that in Base is not yet clear.

It should also be possible to avoid returning a Union{T, Null} vector from predict() when
the input variables are not nullable by storing that in the type information, but
this is left for later.

Also require Julia 0.6.