Move back to DataFrames and port from Nullable to Missings (#30)
This reverts commit 9a5ba5cf9c6eac782494a6fe3eb09b772e2a64a5.

Keep Nullable in cases where it's used to represent "software engineer's nulls",
i.e. not missing values in data, since the replacement for that in Base is not yet ready.

It should also be possible to avoid returning a Union{T, Missing} vector from predict() when
the input variables do not allow for missing values by storing that in the type information, but
this is left for later.

Also require Julia 0.6, and remove useless call to float() which fails with the latest Nulls release.